requirements.txt
about-time==4.2.1
absl-py==2.3.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alive-progress==3.2.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
autograd==1.8.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cma==4.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
deepface==0.0.93
Deprecated==1.2.18
diffusers @ git+https://github.com/huggingface/diffusers.git@1bc6f3dc0f21779480db70a4928d14282c0198ed
dill==0.3.8
executing==2.2.0
facenet-pytorch==2.5.3
fastapi==0.115.12
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
gradio==5.34.0
gradio_client==1.10.3
grapheme==0.6.0
groovy==0.1.2
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.4.4
multiprocess==0.70.16
namex==0.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nncf==2.17.0
numba==0.61.2
numpy==1.26.4
onnx==1.18.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openvino==2025.2.0
openvino-telemetry==2025.1.0
openvino-tokenizers==2025.2.0.0
opt_einsum==3.4.0
optimum==1.26.1
optimum-intel==1.23.0
optree==0.16.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.8
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.2
pymoo==0.6.1.5
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
